refactor(abstract-utxo): factor verifyTransaction
Make room for descriptor-specific implementations

Issue: BTC-1450
